requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose
passlib[bcrypt]
python-dotenv
email-validator

fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
motor
httpx
pytest
pytest-asyncio
